refactor: Utils 폴더 추가 및 util.js 파일 math.js 로 이름 변경
